Use new Permission Sets for permission management
