+++ lib/generators/spree_multi_vendor/install/templates/config/initializers/spree_multi_vendor.rb
@@ -1,0 +1,4 @@
+Rails.application.config.after_initialize do
+  Spree.permissions.assign(:admin, SpreeMultiVendor::PermissionSets::SuperUser)
+  Spree.permissions.assign(:vendor, SpreeMultiVendor::PermissionSets::VendorUser)
+end
